🐛 fix(ui): restore the lazy boundary the modal migration dropped
Replacing the declarative modals with imperative openers turned two dynamic
imports into static ones. Measured against a real SPA build, one of them is a
payload regression and the other is not:

- `ModalProvider` had `lazy(() => import('.../useCreateModal'))`. Made static,
  the create-agent code (and the ChatInput stack behind it) merged into
  `ModalProvider-*.js`, a 32 KB chunk that four `_layout-*.js` chunks import
  statically — so it loaded with every layout mounting AgentModalProvider.
  Moving the import back inside the effect that opens the modal splits it into
  its own 16 KB chunk, now referenced only through mapDeps (dynamic).
- `MessageContent` had `dynamic(() => import('@/features/EditorModal'))`. Made
  static, `@lobehub/editor/react` becomes statically reachable from every
  message row — but that costs nothing today, because the composer already
  pulls the editor into the entry chunk in both builds. Splitting the editor
  half into its own lazily-imported module is kept anyway: it restores the
  boundary the original code expressed and keeps `openEditorModal` free of any
  editor dependency, so the entry chunk can shed the editor later.

Both use `import type` for the remaining type references, since esbuild keeps
`import { type X }` as a side-effect import and that alone would defeat a split.
