+++ src/features/EditorModal/type.ts
@@ -1,0 +1,3 @@
+import type { IEditor } from '@lobehub/editor';
+
+export type EditorRef = { current?: IEditor };
